In device installation instructions, spell out that su is for 14.1
The standard device installation instructions template lists the su
add-on as an optional method for rooting the device. As BUGBASH-240
highlights, it isn't immediately clear that this is for 14.1
only. Thus, for as long as there are LineageOS builds for earlier
versions and the su addon hasn't been backported, make it clear that
the su addon will only work on 14.1.